Version 1.2.21(07/06/2025)
It may be Eid but I'll still do something...
Well I decided to ditch the particles as it would be too time consuming and not worth it, but, there was an issue with level 11, so I decided to fix it! Plus, Evil Robo will also appear even if the lights are off, but he will send you back to the checkpoint. Access denied!